# openapi.yaml
openapi: 3.0.0
info:
  title: TeachReach-API
  description: ASP.NET Core 6.0 microservice for a tutoring marketplace connecting teachers and students
  version: 1.0.0
  contact:
    name: TeachReach Team
    url: https://teachreach.example.com
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.teachreach.example.com
    description: Production server
...
paths:
  /api/teacher/teachers:
    get:
      tags:
        - Teachers
      summary: Get all teachers
      description: Retrieve a list of all teachers with their subjects and reviews
      operationId: getAllTeachers
      responses:
        '200':
          description: List of teachers retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Teacher'
        '500':
          description: Internal server error
  /api/teacher/teacher/{id}:
    get:
      tags:
        - Teachers
      summary: Get teacher by ID
      description: Retrieve a specific teacher with their subjects and reviews
      operationId: getTeacherById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
          description: The teacher ID
      responses:
        '200':
          description: Teacher retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
        '500':
          description: Internal server error
  /api/teacher/teacher/email/{email}:
    get:
      tags:
        - Teachers
      summary: Get teacher by email
      description: Retrieve a specific teacher using their email address
      operationId: getTeacherByEmail
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
          description: The teacher email address
      responses:
        '200':
          description: Teacher retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
        '500':
          description: Internal server error

    post:
      tags:
        - Teachers
      summary: Create a new teacher
      description: Create a new teacher profile
      operationId: createTeacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Teacher'
      responses:
        '200':
          description: Teacher created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveResponse'
        '500':
          description: Internal server error
  /api/teacher/teacher:
    post:
      tags:
        - Teachers
      summary: Create a new teacher
      description: Create a new teacher profile
      operationId: createTeacher2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Teacher'
      responses:
        '200':
          description: Teacher created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveResponse'
        '500':
          description: Internal server error

    put:
      tags:
        - Teachers
      summary: Update a teacher
      description: Update an existing teacher profile
      operationId: updateTeacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
          description: The teacher ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Teacher'
      responses:
        '200':
          description: Teacher updated successfully
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Bad request - ID mismatch
        '500':
          description: Internal server error
  /api/teacher/deleteTeacher/{id}:
    delete:
      tags:
        - Teachers
      summary: Delete a teacher
      description: Delete a teacher profile by ID
      operationId: deleteTeacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
          description: The teacher ID
      responses:
        '200':
          description: Teacher deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveResponse'
        '500':
          description: Internal server error
...
components:
  schemas:
    Teacher:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: Teacher ID
        firstName:
          type: string
          description: Teacher first name
        lastName:
          type: string
          description: Teacher last name
        email:
          type: string
          format: email
          description: Teacher email address
        password:
          type: string
          description: Teacher password (will be hashed)
        bio:
          type: string
          nullable: true
          description: Teacher bio
        city:
          type: string
          nullable: true
          description: Teacher city
        phoneNumber:
          type: string
          description: Teacher phone number
        experience:
          type: integer
          format: int32
          description: Years of teaching experience
        dateOfBirth:
          type: string
          format: date-time
          description: Teacher date of birth
        hourlyRate:
          type: number
          format: decimal
          description: Teacher hourly rate
        profilePictureUrl:
          type: string
          nullable: true
          description: URL to teacher profile picture
        subjects:
          type: array
          items:
            $ref: '#/components/schemas/Subject'
          description: List of subjects taught by teacher
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
          description: List of reviews for teacher
      required:
        - firstName
        - lastName
        - email
        - password
        - phoneNumber
        - dateOfBirth
        - hourlyRate
...
    Subject:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: Subject ID
        name:
          type: string
          description: Subject name
        description:
          type: string
          nullable: true
          description: Subject description
        teachers:
          type: array
          items:
            $ref: '#/components/schemas/Teacher'
          description: Teachers who teach this subject
      required:
        - name
...
    Review:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: Review ID
        teacherId:
          type: integer
          format: int32
          nullable: true
          description: Teacher ID
        studentId:
          type: integer
          format: int32
          nullable: true
          description: Student ID
        rating:
          type: integer
          format: int32
          nullable: true
          minimum: 1
          maximum: 5
          description: Rating from 1 to 5 stars
        comment:
          type: string
          nullable: true
          description: Review comment
        dateCreated:
          type: string
          format: date-time
          description: When the review was created
      required:
        - dateCreated
...
    SaveResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the operation was successful
        errorMessage:
          type: string
          nullable: true
          description: Error message if operation failed